refactor extract useDeleteExtensionaction
